chore: high level restructure of canvas panel files (#69004)
