+++ public/app/plugins/panel/canvas/editor/inline/InlineEdit.tsx
